Version 3.1.0
* Refactoring
* Separation of APIs in client libraries. Currently, supported APIs are:
  * HiroGraph:
    https://core.arago.co/help/specs/?url=definitions/graph.yaml
  * HiroAuth:
    https://core.arago.co/help/specs/?url=definitions/auth.yaml
  * HiroApp:
    https://core.arago.co/help/specs/?url=definitions/app.yaml
