+++ src/hiro_graph_client/VERSION
@@ -1,1 +1,1 @@
-3.0.0
+3.1.0
